requirements.txt
django-fluent-dashboard
django-tables2
django-controlcenter
